update error.log message (#19)
* Remove quotes from input

Input should never have quotes. Inputs are saved in variables, which should not include quotes, otherwise the quotes are added to the path, and includes the current directory, as the input no longer starts with a valid root, as it starts with a quote.
Note: Not tested.

* Fix errors from xml2json

`===` does not work. Reverted to `==`.

* Fix error logging of xml2json

`=== ""` doesn't work. Used `.length === 0` instead. Also changed checking output to checking error, and therefore, `===` to `!==`.

* Remove duplicate first lines in error.log

* update node version

Co-authored-by: Parthiv S Vora <6434386+TheRealPSV@users.noreply.github.com>